WIP refactoring and added form to state
